Move HLSL specific functions to HLSLEmitSource.
