+++ source/slang/slang-emit-cpp.h
@@ -18,6 +18,8 @@ public:
     {}
 
 protected:
+    virtual void emitIRParameterGroupImpl(IRGlobalParam* varDecl, IRUniformParameterGroupType* type) SLANG_OVERRIDE;
+    virtual void emitIREntryPointAttributesImpl(IRFunc* irFunc, EntryPointLayout* entryPointLayout) SLANG_OVERRIDE;
 };
 
 }
