+++ source/slang/slang-emit-glsl.cpp
@@ -1,5 +1,5 @@
-// slang-emit-c-like.cpp
-#include "slang-emit-c-like.h"
+// slang-emit-glsl.cpp
+#include "slang-emit-glsl.h"
 
 #include "../core/slang-writer.h"
 
@@ -10,5 +10,15 @@
 
 namespace Slang {
 
+
+void GLSLSourceEmitter::emitIRParameterGroupImpl(IRGlobalParam* varDecl, IRUniformParameterGroupType* type)
+{
+    emitGLSLParameterGroup(varDecl, type);
+}
+
+void GLSLSourceEmitter::emitIREntryPointAttributesImpl(IRFunc* irFunc, EntryPointLayout* entryPointLayout)
+{
+    emitIREntryPointAttributes_GLSL(irFunc, entryPointLayout);
+}
 
 } // namespace Slang
