+++ source/slang/slang-emit-glsl.h
@@ -19,6 +19,8 @@ public:
 
 protected:
 
+    virtual void emitIRParameterGroupImpl(IRGlobalParam* varDecl, IRUniformParameterGroupType* type) SLANG_OVERRIDE;
+    virtual void emitIREntryPointAttributesImpl(IRFunc* irFunc, EntryPointLayout* entryPointLayout) SLANG_OVERRIDE;
 };
 
 }
